+++ app/controllers/editor/users_controller.rb
@@ -8,8 +8,11 @@ class Editor::UsersController < Editor::ApplicationController
 
   # GET /editor/users/list
   def list
+    options = {}
+    options[:admin] = true if filter_params[:admin].present?
+    options[:editor] = true if filter_params[:editor].present?
     @text = ['username ilike :text or email ilike :text', { text: "%#{filter_params[:text]}%" }] if filter_params[:text].present?
-    @pagy, @users = pagy(User.where(@text))
+    @pagy, @users = pagy(User.where(@text).where(options))
   end
 
   # GET /editor/users/:id
@@ -18,10 +21,14 @@ class Editor::UsersController < Editor::ApplicationController
   # GET /editor/users/:id/tickets
   def tickets
     @text = ['facts.title ilike :string', { string: "%#{filter_params[:text]}%" }] if filter_params[:text].present?
-    @pagy, @tickets = pagy(
-      @user.tickets.joins(happening: [:fact]).where(happening: { fact: @current_user.facts }).where(@text),
-      items: 6
-    )
+    if current_user.admin?
+      @pagy, @tickets = pagy( @user.tickets.joins(happening: [:fact]).where(@text), items: 6)
+    else
+      @pagy, @tickets = pagy(
+        @user.tickets.joins(happening: [:fact]).where(happening: { fact: @current_user.facts }).where(@text),
+        items: 6
+      )
+    end
   end
 
   private
@@ -33,6 +40,6 @@ class Editor::UsersController < Editor::ApplicationController
 
   # Filter params for search a {Fact}
   def filter_params
-    params.fetch(:filter, {}).permit(:text)
+    params.fetch(:filter, {}).permit(:text, :admin, :editor)
   end
 end