+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1]
+### changed
+- fixed fact destroy
+- fixed happening destroy
+- fixed happening creations
+
 ## [2.1.0]
 ### Added
 - Users' ticket manage for editor